requirements.txt
openai
azure-functions
